Async tasks redone

+++ ThirdTask/ThirdTask/Program.cs
@@ -2,6 +2,8 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Xml;
+using System.Xml.Linq;
 using System.Xml.Serialization;
 
 namespace ThirdTask
@@ -10,7 +12,6 @@ namespace ThirdTask
     {
         static void Main(string[] args)
         {
-            Console.WriteLine("Hello World!");
             ConsoleKey key;
             Console.WriteLine("\nPress number. Enter to exit.");
             Console.WriteLine("1.Group by category");
@@ -21,59 +22,68 @@ namespace ThirdTask
             List<Product> products = new();
             List<Product> resProd = new();
 
-            int prodCount = 100;
-            for (int i = 0; i < prodCount; i++)
-            {
-                Product prod = new Product();
-                products.Add(prod);
-            }
 
+            if (!(File.Exists(@"output.xml")))
+            {
+                int prodCount = 100;
+                for (int i = 0; i < prodCount; i++)
+                {
+                    Product product = new();
+                    products.Add(product);
+                }
+                XmlSerializer serialiser = new XmlSerializer(typeof(List<Product>));
+                TextWriter filestream = new StreamWriter(@"output.xml");
+                serialiser.Serialize(filestream, products);
+                filestream.Close();
+            } else
+            {
+                XmlSerializer serializer = new(typeof(List<Product>));
+                StreamReader reader = new(@"output.xml");
+                products = (List<Product>)serializer.Deserialize(reader);
+                reader.Close();
+            }
+                
             do {
                 key = Console.ReadKey(true).Key;
                 if (key == ConsoleKey.D1 || key == ConsoleKey.NumPad1)
                 {
-                    var prodgroups = from product in products group product by product.Category;
+                    var productGroups = from product in products group product by product.category;
 
-                    foreach (IGrouping<Categories, Product> g in prodgroups)
+                    foreach (IGrouping<Categories, Product> group in productGroups)
                     {
-                        Console.Write("\n\n\n" + g.Key + "\n\n\n");
-                        foreach (var t in g)
-                            Console.Write("\n Name: {0}\n Category: {1}\n Price: {2}\n Quantity: {3}\n", t.Name, t.Category, t.Price, t.Quantity);
+                        Console.Write("\n\n\n" + group.Key + "\n\n\n");
+                        foreach (var groupProduct in group)
+                            Console.Write("\n Name: {0}\n Category: {1}\n Price: {2}\n Quantity: {3}\n", groupProduct.name, groupProduct.category, groupProduct.price, groupProduct.quantity);
                     }
                 } else if (key == ConsoleKey.D2 || key == ConsoleKey.NumPad2)
                 {
-                    double qty = 0;
-                    foreach(Product prod in products)
+                    int quantity = 0;
+                    foreach(Product product in products)
                     {
-                        qty += prod.Quantity;
+                        quantity += product.quantity;
                     }
-                    Console.WriteLine("Total Quantity: {0}\n", qty);
-                }
-                else if(key == ConsoleKey.D3 || key == ConsoleKey.NumPad3)
+                    Console.WriteLine("Total Quantity: {0}\n", quantity);
+                } else if(key == ConsoleKey.D3 || key == ConsoleKey.NumPad3)
                 {
                     double sum = 0;
-                    foreach (Product prod in products)
+                    foreach (Product product in products)
                     {
-                        sum += prod.Price * prod.Quantity;
+                        sum += product.price * product.quantity;
                     }
                     Console.WriteLine("Total Price: {0}\n", sum);
-                }
-                else if(key == ConsoleKey.D4 || key == ConsoleKey.NumPad4)
+                } else if(key == ConsoleKey.D4 || key == ConsoleKey.NumPad4)
                 {
-                    var sortprod = from product in products orderby product.Name select product;
+                    var sortprod = from product in products orderby product.name select product;
 
                     foreach (Product prod in sortprod)
                     {
-                        Console.Write("\n Name: {0}\n Category: {1}\n Price: {2}\n Quantity: {3}\n", prod.Name, prod.Category, prod.Price, prod.Quantity);
+                        Console.Write("\n Name: {0}\n Category: {1}\n Price: {2}\n Quantity: {3}\n", prod.name, prod.category, prod.price, prod.quantity);
                     }
                     Console.WriteLine("Sort by name\n");
                 }
             } while (key != ConsoleKey.Enter);
 
-            XmlSerializer serialiser = new XmlSerializer(typeof(List<Product>));
-            TextWriter filestream = new StreamWriter(@"output.xml");
-            serialiser.Serialize(filestream, products);
-            filestream.Close();
+            
         }
     }
 }
